requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
colorama==0.4.6
DateTime==5.2
Django==4.2.6
django-bootstrap-v5==1.0.11
django-filter==23.3
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
iniconfig==2.0.0
Markdown==3.5
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
pytz==2023.3.post1
regex==2023.10.3
setuptools==68.2.2
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
zope.interface==6.1
